requirements.txt
torch==2.0.1
torchaudio==2.0.2
transformers==4.35.2
datasets==2.14.6
pyarrow==14.0.1
accelerate==0.24.1
evaluate==0.4.1
jiwer==3.0.3
soundfile==0.12.1
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.1
librosa==0.10.1
tqdm==4.66.1
pyyaml==6.0.1
scikit-learn==1.3.2
peft==0.6.2
bitsandbytes==0.41.3

